bot/requirements.txt
python-telegram-bot
pymongo
beautifulsoup4
requests

